# Assignments documentation/Assignment_3/swagger.json
{
  "x-generator": "NSwag v13.8.2.0 (NJsonSchema v10.2.1.0 (Newtonsoft.Json v9.0.0.0))",
  "swagger": "2.0",
  "info": {
    "title": "My Title",
    "version": "1.0.0"
  },
  "host": "localhost:5003",
  "schemes": [
    "http"
  ],
  "paths": {
    "/api/Families": {
      "get": {
        "tags": [
          "Families"
        ],
        "operationId": "Families_GetFamilyWithQuery",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "type": "string",
            "name": "streetName",
            "in": "query",
            "x-nullable": true
          },
          {
            "type": "integer",
            "name": "houseNumber",
            "in": "query",
            "format": "int32",
            "x-nullable": true
          },
          {
            "type": "integer",
            "name": "numberOfAdults",
            "in": "query",
            "format": "int32",
            "x-nullable": true
          }
        ],
        "responses": {
          "200": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "$ref": "#/definitions/Family"
            }
          }
        }
      },
      "post": {
        "tags": [
          "Families"
        ],
        "operationId": "Families_AddFamily",
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "name": "family",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Family"
            },
            "x-nullable": false
          }
        ],
        "responses": {
          "200": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "$ref": "#/definitions/Family"
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Families"
        ],
        "operationId": "Families_DeleteFamily",
        "parameters": [
          {
            "type": "string",
            "name": "streetName",
            "in": "query",
            "x-nullable": true
          },
          {
            "type": "integer",
            "name": "houseNumber",
            "in": "query",
            "format": "int32",
            "x-nullable": false
          }
        ],
        "responses": {
          "200": {
            "x-nullable": true,
            "description": "",
            "schema": {
              "type": "file"
            }
          }
        }
      }
    },
    "/api/Families/adults": {
      "delete": {
        "tags": [
          "Families"
        ],
        "operationId": "Families_DeleteAdult",
        "parameters": [
          {
            "type": "integer",
            "name": "id",
            "in": "query",
            "format": "int32",
            "x-nullable": false
          }
        ],
        "responses": {
          "200": {
            "x-nullable": true,
            "description": "",
            "schema": {
              "type": "file"
            }
          }
        }
      }
    },
    "/api/User": {
      "post": {
        "tags": [
          "User"
        ],
        "operationId": "User_Register",
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "name": "user",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/User"
            },
            "x-nullable": false
          }
        ],
        "responses": {
          "200": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        }
      }
    },
    "/api/User/login": {
      "post": {
        "tags": [
          "User"
        ],
        "operationId": "User_Login",
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "type": "string",
            "name": "username",
            "in": "query",
            "x-nullable": true
          },
          {
            "name": "password",
            "in": "body",
            "required": true,
            "schema": {
              "type": "string"
            },
            "x-nullable": false
          }
        ],
        "responses": {
          "200": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        }
      }
    }
  },
  "definitions": {
    "Family": {
      "type": "object",
      "required": [
        "StreetName",
        "HouseNumber"
      ],
      "properties": {
        "StreetName": {
          "type": "string",
          "minLength": 1
        },
        "HouseNumber": {
          "type": "integer",
          "format": "int32",
          "maximum": 999.0,
          "minimum": 1.0
        },
        "Adults": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Adult"
          }
        },
        "Children": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Child"
          }
        },
        "Pets": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Pet"
          }
        }
      }
    },
    "Adult": {
      "allOf": [
        {
          "$ref": "#/definitions/Person"
        },
        {
          "type": "object",
          "required": [
            "JobTitle"
          ],
          "properties": {
            "JobTitle": {
              "type": "string",
              "minLength": 1
            }
          }
        }
      ]
    },
    "Person": {
      "type": "object",
      "required": [
        "Id",
        "FirstName",
        "LastName",
        "HairColor",
        "EyeColor",
        "Age",
        "Weight",
        "Height",
        "Sex"
      ],
      "properties": {
        "Id": {
          "type": "integer",
          "format": "int32"
        },
        "FirstName": {
          "type": "string",
          "maxLength": 40,
          "minLength": 2
        },
        "LastName": {
          "type": "string",
          "maxLength": 40,
          "minLength": 2
        },
        "HairColor": {
          "type": "string",
          "minLength": 1
        },
        "EyeColor": {
          "type": "string",
          "minLength": 1
        },
        "Age": {
          "type": "integer",
          "format": "int32",
          "maximum": 125.0,
          "minimum": 0.0
        },
        "Weight": {
          "type": "number",
          "format": "float",
          "maximum": 250.0,
          "minimum": 1.0
        },
        "Height": {
          "type": "integer",
          "format": "int32",
          "maximum": 300.0,
          "minimum": 30.0
        },
        "Sex": {
          "type": "string",
          "minLength": 1
        }
      }
    },
    "Child": {
      "allOf": [
        {
          "$ref": "#/definitions/Person"
        },
        {
          "type": "object",
          "properties": {
            "ChildInterests": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/ChildInterest"
              }
            },
            "Pets": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Pet"
              }
            }
          }
        }
      ]
    },
    "ChildInterest": {
      "type": "object",
      "required": [
        "childId"
      ],
      "properties": {
        "childId": {
          "type": "integer",
          "format": "int32"
        },
        "interestId": {
          "type": "string"
        }
      }
    },
    "Pet": {
      "type": "object",
      "required": [
        "id",
        "species",
        "name",
        "age"
      ],
      "properties": {
        "id": {
          "type": "integer",
          "format": "int32"
        },
        "species": {
          "type": "string",
          "minLength": 1
        },
        "name": {
          "type": "string",
          "minLength": 1
        },
        "age": {
          "type": "integer",
          "format": "int32"
        }
      }
    },
    "User": {
      "type": "object",
      "required": [
        "Username",
        "Password",
        "Department",
        "SecurityLevel"
      ],
      "properties": {
        "Username": {
          "type": "string",
          "maxLength": 25,
          "minLength": 5
        },
        "Password": {
          "type": "string",
          "maxLength": 25,
          "minLength": 5
        },
        "Department": {
          "type": "string",
          "minLength": 1
        },
        "SecurityLevel": {
          "type": "integer",
          "format": "int32"
        }
      }
    }
  }
}