# DNP_Assignment_1/Assignment_1/Models/Child.cs
using System.Collections.Generic;
using System.Text.Json.Serialization;

namespace Models {
    public class Child : Person
    {
        [JsonPropertyName("ChildInterests")]
        public List<ChildInterest> ChildInterests { get; set; }
        [JsonPropertyName("Pets")]
        public List<Pet> Pets { get; set; }

        public void Update(Child toUpdate)
        {
            base.Update(toUpdate);
            ChildInterests = toUpdate.ChildInterests;
            Pets = toUpdate.Pets;
        }

    }
}