# WebAPI/FamilyWebAPI/Models/Adult.cs
using System.ComponentModel.DataAnnotations;
using System.Text.Json;
using System.Text.Json.Serialization;

namespace Models
{
    public class Adult : Person
    {
        [Required(ErrorMessage = "Enter a job position"), JsonPropertyName("JobTitle")]

        public string JobTitle { get; set; }

        public override string ToString()
        {
            return JsonSerializer.Serialize(this);
        }

        public void Update(Adult toUpdate)
        {
            JobTitle = toUpdate.JobTitle;
            base.Update(toUpdate);
        }

    }
}